visual_planner/requirements.txt
huggingface-hub
diffusers
matplotlib==3.7.5
einops==0.7.0
einops-exts==0.0.4
ema-pytorch==0.2.3
tqdm==4.66.1
accelerate==0.23.0
transformers==4.34.0
pytorch-fid==0.3.0
pynvml==11.5.0
tensorboard
imageio[ffmpeg]
wandb==0.16.6
opencv-python==4.9.0.80
pytorch-lightning==1.8.6
rotary-embedding-torch==0.5.3

